+++ internal/clientset/clientset.go
@@ -53,7 +53,7 @@ type ClientSet struct {
 	legacySlos     *cxsdk.LegacySLOsClient
 	dashboards     *cxsdk.DashboardsClient
 	ruleGroups     *cxsdk.RuleGroupsClient
-	users          *cxsdk.UsersClient
+	users          *UsersClient
 	events2Metrics *cxsdk.Events2MetricsClient
 	groupGrpc      *cxsdk.GroupsClient
 	teams          *cxsdk.TeamsClient
@@ -167,7 +167,7 @@ func (c *ClientSet) Groups() *GroupsClient {
 	return c.groups
 }
 
-func (c *ClientSet) Users() *cxsdk.UsersClient {
+func (c *ClientSet) Users() *UsersClient {
 	return c.users
 }
 
@@ -204,9 +204,9 @@ func (c *ClientSet) DataEnrichments() (*ess.EnrichmentsServiceAPIService, *cess.
 	return c.dataEnrichments, c.customDataEnrichments
 }
 
-func NewClientSet(region string, apiKey string, targetUrl string) *ClientSet {
-	apiKeySdk := cxsdk.NewSDKCallPropertiesCreatorTerraform(strings.ToLower(region), cxsdk.NewAuthContext(apiKey, apiKey), TF_PROVIDER_VERSION)
-	apikeyCPC := NewCallPropertiesCreator(targetUrl, apiKey)
+func NewClientSet(region string, apiKey string, grpcTarget string) *ClientSet {
+	grpcCreator := newTerraformSDKCallPropertiesCreator(apiKey, TF_PROVIDER_VERSION, grpcTarget)
+	apikeyCPC := NewCallPropertiesCreator(grpcTarget, apiKey)
 
 	confBuilder := cxsdkOpenapi.NewConfigBuilder().
 		WithTerraformVersion(TF_PROVIDER_VERSION).
@@ -231,18 +231,18 @@ func NewClientSet(region string, apiKey string, targetUrl string) *ClientSet {
 
 	return &ClientSet{
 		// deprecated
-		dataSet:     cxsdk.NewDataSetClient(apiKeySdk),
-		enrichments: cxsdk.NewEnrichmentClient(apiKeySdk),
-		legacySlos:  cxsdk.NewLegacySLOsClient(apiKeySdk),
-		ruleGroups:  cxsdk.NewRuleGroupsClient(apiKeySdk),
-		teams:       cxsdk.NewTeamsClient(apiKeySdk),
+		dataSet:     cxsdk.NewDataSetClient(grpcCreator),
+		enrichments: cxsdk.NewEnrichmentClient(grpcCreator),
+		legacySlos:  cxsdk.NewLegacySLOsClient(grpcCreator),
+		ruleGroups:  cxsdk.NewRuleGroupsClient(grpcCreator),
+		teams:       cxsdk.NewTeamsClient(grpcCreator),
 
-		users: cxsdk.NewUsersClient(apiKeySdk),
+		users: NewUsersClient(region, apiKey),
 
 		// TODO
-		dashboards:     cxsdk.NewDashboardsClient(apiKeySdk),
-		events2Metrics: cxsdk.NewEvents2MetricsClient(apiKeySdk),
-		groupGrpc:      cxsdk.NewGroupsClient(apiKeySdk),
+		dashboards:     cxsdk.NewDashboardsClient(grpcCreator),
+		events2Metrics: cxsdk.NewEvents2MetricsClient(grpcCreator),
+		groupGrpc:      cxsdk.NewGroupsClient(grpcCreator),
 
 		dahboardsFolders:      cs.DashboardFolders(),
 		parsingRuleGroups:     cs.RuleGroups(),
@@ -267,6 +267,6 @@ func NewClientSet(region string, apiKey string, targetUrl string) *ClientSet {
 		customDataEnrichments: cs.CustomEnrichments(),
 		alertScheduler:        cs.AlertScheduler(),
 		grafana:               NewGrafanaClient(apikeyCPC),
-		groups:                NewGroupsClient(apikeyCPC),
+		groups:                NewGroupsClient(region, apiKey),
 	}
 }